Partial Update to perl6 Spec v1.0
Now compiles and runs, though slow.
Seems to now output correct values, however, our list seem to grow
fatter as we increase the count. More explicit flattening is needed
somewhere.
Increasing count also creates extra output... Count is suppouse to
return N number of jobs which are next to be executed. It is possible
for more than 1 job to be listed with Count 1, but that should only
happen if 2 jobs are set to be run at the same time.
In other words, Count sets the number of next cronjobtimes that will be
returned, the number of jobs associated with those times is variable.