Fix NextJob range slicing
Rakudo does not yet Trancat Range Slices to only include ranges within
the list. While this feature is documented, it does not show up in the
specs nor in the official test suit, so this may be the intended result.

This commit alters the slice range to ensure we always get defined
elements